+++ src/screens/home/components/transactions/hooks.tsx
@@ -1,43 +1,36 @@
 import { useState } from 'react';
 import {
-  TransactionsState, TransactionType,
-} from './types';
-
-const fakeData:TransactionType = {
-  slot: 812768640,
-  success: true,
-  messages: 5,
-  signature: '76nwV8zz8tLz97SBRXH6uwHvgHXtqJDLQfF66jZhQ857',
-  timestamp: '2021-09-13T20:06:17.363145',
-};
+  useTransactionsListenerSubscription,
+  TransactionsListenerSubscription,
+} from '@graphql/types';
+import { TransactionsState } from './types';
 
 export const useTransactions = () => {
-  const [state, _setState] = useState<TransactionsState>({
-    items: Array(7).fill(fakeData),
+  const [state, setState] = useState<TransactionsState>({
+    items: [],
   });
 
   // ================================
   // txs subscription
   // ================================
-  // useTransactionsListenerSubscription({
-  //   onSubscriptionData: (data) => {
-  //     setState({
-  //       items: formatTransactions(data.subscriptionData.data),
-  //     });
-  //   },
-  // });
+  useTransactionsListenerSubscription({
+    onSubscriptionData: (data) => {
+      setState({
+        items: formatTransactions(data.subscriptionData.data),
+      });
+    },
+  });
 
-  // const formatTransactions = (data: TransactionsListenerSubscription) => {
-  //   return data.transactions.map((x) => {
-  //     return ({
-  //       height: x.height,
-  //       hash: x.hash,
-  //       success: x.success,
-  //       timestamp: x.block.timestamp,
-  //       messages: x.messages.length,
-  //     });
-  //   });
-  // };
+  const formatTransactions = (data: TransactionsListenerSubscription) => {
+    return data.transactions.map((x) => {
+      return ({
+        slot: x.slot,
+        hash: x.hash,
+        success: !x.error,
+        timestamp: x.block.timestamp,
+      });
+    });
+  };
 
   return {
     state,