+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Changes
 - Add latest blocks (home) ([\#16](https://github.com/forbole/big-dipper-2.0-solana/issues/16))
 - Add average slot time (home) ([\#14](https://github.com/forbole/big-dipper-2.0-solana/issues/14))
+- Add latest transactions (home) ([\#17](https://github.com/forbole/big-dipper-2.0-solana/issues/17))